Update dimr_dflowfm_waq_coupling_1D2D.xml (#344)
* Update dimr_dflowfm_waq_coupling_1D2D.xml

New timestamp and new test case

* Update the reference files for the PROTIST test cases

As we found a small but important bug in the PROTIST code (sedimentation of diatoms), the original reference files for the related cases are no longer valid.

* DELWAQ-1109: Revert the changes to the timestamps

Because of the inconsistency in the timestamps for a numbe of PROTIST test cases, I have reverted the change. Hopefully this will solve the issue that 10 unrelated cases are failing (at least for now)

---------

Co-authored-by: Qinghua_ye <qinghua.ye@deltares.nl>
Co-authored-by: Qinghua Ye <113536652+ye4deltares@users.noreply.github.com>